+++ packages/nimble-components/src/rich-text/editor/tests/rich-text-editor.stories.ts
@@ -13,12 +13,16 @@ import {
 } from '../../../label-provider/base/tests/label-user-stories-utils';
 import { labelProviderRichTextTag } from '../../../label-provider/rich-text';
 import { richTextMarkdownString } from '../../../utilities/tests/rich-text-markdown-string';
+import { mappingUserTag } from '../../../mapping/user';
+import { richTextMentionUsersTag } from '../../../rich-text-mention/users';
 
 // eslint-disable-next-line @typescript-eslint/no-empty-interface
 interface RichTextEditorArgs extends LabelUserArgs {
     data: ExampleDataType;
+    mentionData: MentionDataType;
     footerActionButtons: boolean;
     getMarkdown: undefined;
+    getMentionedHrefs: undefined;
     editorRef: RichTextEditor;
     setMarkdownData: (args: RichTextEditorArgs) => void;
     disabled: boolean;
@@ -44,6 +48,21 @@ const dataSets = {
     [exampleDataType.markdownString]: richTextMarkdownString
 } as const;
 
+type MentionDataType = (typeof mentionDataType)[keyof typeof mentionDataType];
+
+const mentionDataType = {
+    userPattern: 'UserPattern',
+    httpsPattern: 'HttpsPattern'
+} as const;
+
+const mentionDataSets = {
+    [mentionDataType.userPattern]: { pattern: '^user:(.*)', href: 'user:' },
+    [mentionDataType.httpsPattern]: {
+        pattern: '^https://user/(.*)',
+        href: 'https://user/'
+    }
+} as const;
+
 const richTextEditorDescription = 'The rich text editor component allows users to add/edit text formatted with various styling options including bold, italics, numbered lists, and bulleted lists. The editor generates markdown output and takes markdown as input. The markdown flavor used is [CommonMark](https://spec.commonmark.org/0.30/).\n\n See the [rich text viewer](?path=/docs/incubating-rich-text-viewer--docs) component to render markdown without allowing editing.';
 const setMarkdownDescription = 'A function that sets content in the editor with the provided markdown string.';
 const getMarkdownDescription = 'A function that serializes the current data in the editor and returns the markdown string.';
@@ -56,7 +75,6 @@ client application must implement that functionality.
 
 const metadata: Meta<RichTextEditorArgs> = {
     title: 'Incubating/Rich Text Editor',
-    tags: ['autodocs'],
     decorators: [withActions],
     parameters: {
         docs: {
@@ -65,7 +83,7 @@ const metadata: Meta<RichTextEditorArgs> = {
             }
         },
         actions: {
-            handles: ['input']
+            handles: ['input', 'mention-update']
         }
     },
     // prettier-ignore
@@ -76,7 +94,7 @@ const metadata: Meta<RichTextEditorArgs> = {
     })}
     <${richTextEditorTag}
         ${ref('editorRef')}
-        style="height: 160px;"
+        style="height: 300px;"
         data-unused="${x => x.setMarkdownData(x)}"
         ?disabled="${x => x.disabled}"
         ?footer-hidden="${x => x.footerHidden}"
@@ -84,6 +102,13 @@ const metadata: Meta<RichTextEditorArgs> = {
         error-text="${x => x.errorText}"
         placeholder="${x => x.placeholder}"
     >
+        <${richTextMentionUsersTag} pattern="${x => mentionDataSets[x.mentionData].pattern}">
+            <${mappingUserTag} key="${x => mentionDataSets[x.mentionData].href}1" display-name="John Doe"></${mappingUserTag}>
+            <${mappingUserTag} key="${x => mentionDataSets[x.mentionData].href}2" display-name="Mary Wilson"></${mappingUserTag}>
+            <${mappingUserTag} key="${x => mentionDataSets[x.mentionData].href}3" display-name="Sue Ann"></${mappingUserTag}>
+            <${mappingUserTag} key="${x => mentionDataSets[x.mentionData].href}4" display-name="Joseph George"></${mappingUserTag}>
+            <${mappingUserTag} key="${x => mentionDataSets[x.mentionData].href}5" display-name="David"></${mappingUserTag}>
+        </${richTextMentionUsersTag}>
         ${when(x => x.footerActionButtons, html`
             <${buttonTag} appearance="ghost" slot="footer-actions">Cancel</${buttonTag}>
             <${buttonTag} slot="footer-actions">OK</${buttonTag}>`)}
@@ -103,12 +128,32 @@ const metadata: Meta<RichTextEditorArgs> = {
                 }
             }
         },
+        mentionData: {
+            name: '@mention configuration',
+            description:
+                'Configure how mentions are detected and displayed. See documentation of the `pattern` attribute of the mention configuration element and the `key` attribute of mapping element(s).',
+            options: Object.values(mentionDataType),
+            control: {
+                type: 'radio',
+                labels: {
+                    [mentionDataType.userPattern]: 'User Pattern - user:(.*)',
+                    [mentionDataType.httpsPattern]:
+                        'HTTPS Pattern - https://user/(.*)'
+                }
+            }
+        },
         footerActionButtons: {
             description: footerActionButtonDescription
         },
         getMarkdown: {
             name: 'getMarkdown()',
             description: getMarkdownDescription,
+            control: false
+        },
+        getMentionedHrefs: {
+            name: 'getMentionedHrefs()',
+            description:
+                'Returns an array of strings listing the hrefs of current mentions in the rich text components.',
             control: false
         },
         editorRef: {
@@ -146,7 +191,8 @@ const metadata: Meta<RichTextEditorArgs> = {
         }
     },
     args: {
-        data: exampleDataType.plainString,
+        data: exampleDataType.markdownString,
+        mentionData: mentionDataType.userPattern,
         footerActionButtons: false,
         disabled: false,
         footerHidden: false,